ckpem: make pem_objs a list instead of array
Since libcurl started to use PK11_CreateManagedGenericObject(),
the calls to PK11_DestroyGenericObject() started to take affect
at the level of nss-pem.  Internal objects were removed on each
call of curl_easy_cleanup() and allocated again on each call of
curl_easy_perform().  This would be fine if the corresponding items
of the global pem_objects array were reused, which was, however,
not the case.  Consequently, the array grew unboundedly while most
of its items were NULL pointers.  This caused severe performance
regression in libcurl-based applications.

This commit replaces the global array by a global linked list,
which allows to delete arbitrary items instead of keeping NULL
pointers.  For compatibility reasons, the original array indexes
are stored in the list nodes.

Bug: https://bugzilla.redhat.com/1659108