Fix buffer overrun in ReadDERFromFile
This commit fixes an issue in ReadDERFromFile where it may overrun
the buffer containing the certificate bundle file when searching
for embedded certificates. In some cases, remnants of a previously
read certificate may be found past the buffer resulting in an error
improperly being returned.

The issue is that ReadDERFromFile solely relies on strstr returning
a nullptr when searching for the beginning of the next certificate.
Without a null terminator strstr is free search beyond the intended
buffer. The issue was readily seen by making a curl request in a loop.

The fix is add a null terminator at the end of the read buffer. A
check is also added that we are not beyond the end of the buffer.

Closes: https://github.com/kdudka/nss-pem/pull/10