psession: find the key to decrypt in reverse order
If we attempt to decrypt a key that is already decrypted, curl fails
with a misleading error: `Unable to load client key: Incorrect
password`.

In practice, we usually want to decrypt the key that was loaded the last
time.  Reversing the order of search through the global array makes
nss-pem work in a scenario where 2 distinct encrypted private keys are
used.

Bug: https://bugzilla.redhat.com/2121064